+++ src/main/java/just/plots/commands/ResetCommand.java
@@ -2,6 +2,7 @@ package just.plots.commands;
 
 import just.plots.JustPlots;
 import just.plots.Plot;
+import just.plots.events.PlotClearEvent;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
@@ -42,6 +43,14 @@ public class ResetCommand extends SubCommand {
         if (!senderUuid.equals(plot.getOwner()) && !sender.hasPermission("justplots.reset.other")) {
             sender.sendMessage(ChatColor.RED + JustPlots.getUsername(plot.getOwner()) + " owns that plot");
             return false;
+        }
+
+        PlotClearEvent event = new PlotClearEvent(plot, (Player) sender, true);
+
+        Bukkit.getServer().getPluginManager().callEvent(event);
+
+        if (event.isCancelled()) {
+            throw new RuntimeException("Event was cancelled");
         }
 
         plot.delete();
